pyproject.toml
[tool.poetry]
name = "egrassa_gendiff"
version = "0.6.0"
description = "my project lvl 2"
authors = ["Egrassa <bacek90@mail.ru>"]
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Natural Language :: Russian",
	"Programming Language :: Python :: 3.6"
]
repository = "https://github.com/xegrassa/python-project-lvl2"
packages = [
    { include = "gendiff" },
  ]


[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
pytest = "^5.3"
coverage = "^4.5"
pytest-cov = "^2.8"

[tool.poetry.scripts]
gendiff = "gendiff.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
